change oninit methods
